requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.3.1
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.2
Django==3.2
django-ajax-datatable==4.1.7
django-convenient-formsets==0.9.2
django-cors-headers==3.5.0
django-crispy-forms==1.11.2
django-datatable-view==2.1.5
django-extra-views==0.13.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-durin==0.3.0
django-rest-knox==4.1.0
django-reversetag==0.3.2
django-static-jquery3==5.1.0
django-templated-mail==1.1.1
django-view-breadcrumbs==2.1.1
djangorestframework==3.12.1
djangorestframework-datatables==0.5.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-extra-fields==3.0.3
humanize==3.9.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
livereload==2.6.3
Markdown==3.3.3
MarkupSafe==2.0.1
mysqlclient==2.0.1
oauthlib==3.1.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
toml==0.10.2
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.6
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.3.1
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.2
Django==3.2
django-ajax-datatable==4.1.7
django-convenient-formsets==0.9.2
django-cors-headers==3.5.0
django-crispy-forms==1.11.2
django-datatable-view==2.1.5
django-extra-views==0.13.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-durin==0.3.0
django-rest-knox==4.1.0
django-reversetag==0.3.2
django-static-jquery3==5.1.0
django-templated-mail==1.1.1
django-view-breadcrumbs==2.1.1
djangorestframework==3.12.1
djangorestframework-datatables==0.5.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-extra-fields==3.0.3
humanize==3.9.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
livereload==2.6.3
Markdown==3.3.3
MarkupSafe==2.0.1
mysqlclient==2.0.1
oauthlib==3.1.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
toml==0.10.2
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.6
